backend/user-service/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
bcrypt==4.1.2
PyJWT==2.8.0
pymongo==4.7.1
python-dotenv==1.0.1